move websocket handler to own file
